# api-docs/openapi/openapi.base.yaml
openapi: 3.1.0
info:
  title: XHS Interface Layer API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
  description: |
    API-first contract for the XHS interface layer.

    The interface layer exposes API-only atomic tools over isolated sessions. Each
    session owns its own user data directory. A session must be bound before any
    state-changing or crawling operation can run, and the returned token is the
    proof of current ownership.
servers:
  - url: http://localhost:8000
    description: Local development server
security: []
...
paths:
...
  /v1/sessions/{session_id}/login:
    post:
      tags: [Login]
      summary: Run login action
      description: Uses one endpoint for all login actions. The action field selects the concrete login behavior.
      operationId: runLoginAction
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginActionRequest"
            examples:
              qrcode:
                value:
                  token: "btok_example"
                  action: qrcode
              cookie:
                value:
                  token: "btok_example"
                  action: cookie
                  cookie: "a1=...; web_session=..."
              phone_start:
                value:
                  token: "btok_example"
                  action: phone_start
                  phone: "+8613800000000"
              phone_submit:
                value:
                  token: "btok_example"
                  action: phone_submit
                  code: "123456"
      responses:
        "200":
          description: Login action accepted or completed.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/LoginStatus"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/InvalidSessionToken"
        "404":
          $ref: "#/components/responses/SessionNotFound"
        "429":
          $ref: "#/components/responses/RiskBlocked"
        "503":
          $ref: "#/components/responses/BackendUnavailable"
  /v1/sessions/{session_id}/login/status:
    get:
      tags: [Login]
      summary: Get login status
      description: Pure query. Token is optional and only used for logging and traceability when present.
      operationId: getLoginStatus
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/OptionalToken"
      responses:
        "200":
          description: Login status.
          content:
            application/json:
              examples:
                qrcode:
                  externalValue: ./examples/login-status-qrcode-response.json
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/LoginStatus"
        "404":
          $ref: "#/components/responses/SessionNotFound"
        "503":
          $ref: "#/components/responses/BackendUnavailable"
...
components:
  parameters:
    SessionId:
      name: session_id
      in: path
      required: true
      schema:
        type: string
      example: sess_01hv7r6s9z8x
    OptionalToken:
      name: token
      in: query
      required: false
      schema:
        type: string
      example: btok_example
...
  responses:
    InvalidRequest:
      description: "Error code: invalid_request. Possible causes: request body, path, or query parameters do not match this operation's schema. Action: fix_request."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    InvalidSessionToken:
      description: "Error code: invalid_session_token. Possible causes: token is missing, malformed, expired by unbind, or does not own this session. Action: bind_session."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    SessionNotFound:
      description: "Error code: session_not_found. Possible causes: session_id does not exist or was destroyed. Action: fix_request."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    RiskBlocked:
      description: "Error code: risk_blocked. Possible causes: platform risk control, verification requirement, or network/IP trust issue. Action: inspect_diagnostics."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    BackendUnavailable:
      description: "Error code: backend_unavailable. Possible causes: browser, storage, queue, or internal dependency is unavailable. Action: wait_and_retry."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    LoginActionRequest:
      oneOf:
        - $ref: "#/components/schemas/QrcodeLoginActionRequest"
        - $ref: "#/components/schemas/CookieLoginActionRequest"
        - $ref: "#/components/schemas/PhoneStartLoginActionRequest"
        - $ref: "#/components/schemas/PhoneSubmitLoginActionRequest"
      discriminator:
        propertyName: action
        mapping:
          qrcode: "#/components/schemas/QrcodeLoginActionRequest"
          cookie: "#/components/schemas/CookieLoginActionRequest"
          phone_start: "#/components/schemas/PhoneStartLoginActionRequest"
          phone_submit: "#/components/schemas/PhoneSubmitLoginActionRequest"
    QrcodeLoginActionRequest:
      type: object
      required: [token, action]
      properties:
        token:
          type: string
        action:
          type: string
          const: qrcode
    CookieLoginActionRequest:
      type: object
      required: [token, action, cookie]
      properties:
        token:
          type: string
        action:
          type: string
          const: cookie
        cookie:
          type: string
          description: Required when action is cookie.
    PhoneStartLoginActionRequest:
      type: object
      required: [token, action, phone]
      properties:
        token:
          type: string
        action:
          type: string
          const: phone_start
        phone:
          type: string
          description: Required when action is phone_start.
    PhoneSubmitLoginActionRequest:
      type: object
      required: [token, action, code]
      properties:
        token:
          type: string
        action:
          type: string
          const: phone_submit
        code:
          type: string
          description: Required when action is phone_submit.
...
    SuccessEnvelope:
      type: object
      required: [ok, data, meta]
      properties:
        ok:
          type: boolean
          const: true
        data:
          type: object
          additionalProperties: true
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ErrorEnvelope:
      type: object
      required: [ok, error, meta]
      properties:
        ok:
          type: boolean
          const: false
        error:
          $ref: "#/components/schemas/ApiError"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ResponseMeta:
      type: object
      required: [request_id, operation, retry_count, warnings]
      description: Shared response metadata. Required fields are returned by every operation. Optional fields are omitted when they do not apply; the service does not return null placeholders.
      properties:
        request_id:
          type: string
          description: Unique identifier for this interface-layer request. Use it to correlate client errors with server logs.
        session_id:
          type: string
          description: Returned only when the operation is scoped to a specific session.
        operation:
          type: string
          description: Stable operation identifier, matching the OpenAPI operationId.
        retry_count:
          type: integer
          description: Number of internal retries performed by the interface layer before returning this response.
          default: 0
        warnings:
          type: array
          description: Non-fatal warnings produced while handling this request.
          items:
            type: string
        diagnostics_ref:
          type: string
          description: Returned only when detailed diagnostics can be queried with the diagnostics endpoint.
    ApiError:
      type: object
      required: [code, message, retryable, action]
      description: Stable machine-readable error object. Client code should branch on code, retryable, and action instead of parsing message text.
      properties:
        code:
          type: string
          description: Stable business error code. This is the primary field for client-side branching.
          enum:
            - invalid_request
            - session_not_found
            - session_busy
            - session_quota_exceeded
            - session_create_failed
            - session_destroy_failed
            - invalid_session_token
            - invalid_context
            - stale_context
            - login_required
            - risk_blocked
            - invalid_xsec_token
            - note_unavailable
            - diagnostics_not_found
            - admin_unauthorized
            - request_failed
            - parser_changed
            - backend_unavailable
        message:
          type: string
          description: Human-readable error summary. It is not a stable contract for program logic.
        retryable:
          type: boolean
          description: Whether the exact same request can be retried without changing parameters or session context.
        action:
          type: string
          description: Suggested next action for the caller. This is a stable enum for application-layer recovery flow.
          enum:
            - fix_request
            - bind_session
            - unbind_or_force_unbind
            - login
            - reopen_note
            - refresh_search
            - wait_and_retry
            - inspect_diagnostics
            - contact_admin
        details:
          type: object
          description: Safe structured context for debugging. Sensitive tokens, cookies, and raw private payloads must not be returned here.
          additionalProperties: true
...
    LoginStatus:
      type: object
      description: Current login flow state. QR-code login can require multiple sequential challenges; callers should always display the latest challenge_id/sequence returned here.
      properties:
        state:
          type: string
          enum: [not_started, waiting_qr_scan, waiting_qr_confirm, waiting_sms, ready, failed, risk_blocked, login_required]
        required_action:
          type: string
          enum: [none, scan_qr, confirm_qr, submit_sms_code, retry_login, inspect_diagnostics, cookie]
          default: none
        challenge:
          $ref: "#/components/schemas/LoginChallenge"
        login_user:
          type: object
          additionalProperties: true
        red_id:
          type: string
        user_id:
          type: string
        nickname:
          type: string
        last_error:
          $ref: "#/components/schemas/ApiError"
    LoginChallenge:
      type: object
      required: [challenge_id, type, sequence, status]
      properties:
        challenge_id:
          type: string
          description: Stable identifier for the current login challenge. A new value means the application should refresh its displayed challenge.
        type:
          type: string
          enum: [login_qr, security_qr, sms_code]
          description: Challenge kind. security_qr covers secondary QR verification after an initial QR scan succeeds.
        sequence:
          type: integer
          minimum: 1
          description: Monotonic challenge sequence within the current login flow.
        status:
          type: string
          enum: [pending, scanned, confirmed, expired]
        image_base64:
          type: string
          description: QR image payload when the current challenge requires scanning.
        image_mime:
          type: string
          default: image/png
        expires_at:
          type: string
          format: date-time
        issued_at:
          type: string
          format: date-time
        message:
          type: string
...
  securitySchemes:
    AdminApiKey:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: Admin API key.